+++ Plater/CHANGELOG.md
@@ -1,4 +1,7 @@
 @cont1nuity:
-- Fixing healthbar animation.
-- Disable garbage collection profiling in preparation for 11.1.5.
+- Changing Drain Soul execute for MoP classic.
+- Fixing an error in quest tracking for MoP for certain units.
+- Changing SW:D execute for MoP classic.
+- Adding 'Soul Reaper' as DK execute for MoP classic.
+- Enabling absorb shields when available (MoP classic prep).